api/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
click==8.1.7
fastapi==0.109.2
h11==0.14.0
idna==3.6
pydantic==2.6.0
pydantic_core==2.16.1
redis==5.0.1
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.0.post1
Flask==3.0.3
PyYAML>=6.0
requests==2.32.3
redisearch==2.0.0


